[Refactor] AppTitleInfoView 하위 뷰로 분리
- 앱 타이틀 영역 하위 뷰로 분리
