+++ 35-Seminar/Week2/ViewControllers/AppstoreDetailController.swift
@@ -16,55 +16,10 @@ final class AppstoreDetailController: UIViewController {
     
     private let containerStackView = UIStackView().then {
         $0.axis = .vertical
+        $0.distribution = .fill
     }
     
-    private let titleInfoView = UIView()
-    
-    private let titleLabelContainerStackView = UIStackView().then {
-        $0.axis = .vertical
-        $0.spacing = 4
-    }
-    
-    private let appIconImageView = UIImageView().then {
-        $0.image = UIImage(named: "toss_icon")
-        $0.contentMode = .scaleAspectFit
-        $0.clipsToBounds = true
-        $0.layer.cornerRadius = 25
-        $0.layer.borderWidth = 1
-        $0.layer.borderColor = UIColor.Week2ColorSet.logoImageBorderColor.cgColor
-    }
-    
-    private let titleLabel = UILabel().then {
-        $0.attributedText = .makeAttributedString(text: "토스",
-                                                  color: .black,
-                                                  font: UIFont.systemFont(ofSize: 22, weight: .semibold))
-    }
-    
-    private let descriptionLabel = UILabel().then {
-        $0.attributedText = .makeAttributedString(text: "금융이 쉬워진다.",
-                                                  color: .lightGray,
-                                                  font: UIFont.systemFont(ofSize: 15, weight: .medium))
-    }
-    
-    private let downloadButton = UIButton().then {
-        $0.backgroundColor = UIColor.Week2ColorSet.primaryBlue
-        $0.setTitle("열기", for: .normal)
-        $0.setTitleColor(.white, for: .normal)
-        $0.setAttributedTitle(.makeAttributedString(text: "열기",
-                                                    color: .white,
-                                                    font: UIFont.systemFont(ofSize: 15, weight: .semibold)),
-                              for: .normal)
-        $0.clipsToBounds = true
-        $0.layer.cornerRadius = 14
-    }
-    
-    private let shareIconButton = UIButton().then {
-        let imageConfig = UIImage.SymbolConfiguration(pointSize: 20, weight: .semibold)
-        let image = UIImage(systemName: "square.and.arrow.up", withConfiguration: imageConfig)
-        $0.setImage(image, for: .normal)
-        $0.tintColor = .systemBlue
-        $0.contentMode = .scaleAspectFit
-    }
+    private let appTitleInfoView = AppTitleInfoView()
     
     private let appInfoHorizontalSwipeContainerView = UIView()
     
@@ -608,16 +563,8 @@ final class AppstoreDetailController: UIViewController {
         view.addSubViews(
             containerScrollView.addSubViews(
                 containerStackView.addArrangedSubViews(
-                    // 타이틀 영역
-                    titleInfoView.addSubViews(
-                        appIconImageView,
-                        titleLabelContainerStackView.addArrangedSubViews(
-                            titleLabel,
-                            descriptionLabel
-                        ),
-                        downloadButton,
-                        shareIconButton
-                    ),
+                    // 앱 타이틀 정보 뷰
+                    appTitleInfoView,
                     // 앱 정보 수평 스크롤 뷰
                     appInfoHorizontalSwipeContainerView.addSubViews(
                         appInfoHorizontalSwipeStackView.addArrangedSubViews(
@@ -731,28 +678,7 @@ final class AppstoreDetailController: UIViewController {
             $0.width.equalToSuperview()
         }
         
-        appIconImageView.snp.makeConstraints {
-            $0.top.bottom.left.equalToSuperview().inset(20)
-            $0.size.equalTo(120)
-        }
-        
-        titleLabelContainerStackView.snp.makeConstraints {
-            $0.top.equalTo(appIconImageView.snp.top)
-            $0.left.equalTo(appIconImageView.snp.right).offset(15)
-            $0.right.equalToSuperview().inset(20)
-        }
-        
-        downloadButton.snp.makeConstraints {
-            $0.left.equalTo(appIconImageView.snp.right).offset(15)
-            $0.bottom.equalTo(appIconImageView.snp.bottom)
-            $0.height.equalTo(28)
-            $0.width.equalTo(68)
-        }
-        
-        shareIconButton.snp.makeConstraints {
-            $0.right.equalToSuperview().inset(20)
-            $0.bottom.equalTo(appIconImageView.snp.bottom)
-        }
+        containerStackView.setCustomSpacing(20, after: appTitleInfoView)
         
         // MARK: 수평 스크롤 뷰
         appInfoHorizontalSwipeContainerView.snp.makeConstraints {
